+++ NeuralNetwork/NetworkTraining.h
@@ -5,9 +5,9 @@
 #ifndef OCR_NETWORKTRAINING_H
 #define OCR_NETWORKTRAINING_H
 
-float * feed_forward(Network * network, const float * input_data);
+double * feed_forward(Network * network, const double * input_data);
 void train_neural_network(Network * network,
-                          float ** input_data, float ** expected_output, size_t data_size,
-                          size_t epochs, float eta);
+                          double ** input_data, double ** expected_output, size_t data_size,
+                          size_t epochs, double eta);
 
 #endif //OCR_NETWORKTRAINING_H
